refactor SQLExecutionEvent
