+++ sharding-jdbc-transaction-parent/sharding-jdbc-transaction/src/main/java/io/shardingsphere/transaction/bed/sync/BestEffortsDeliveryListener.java
@@ -19,7 +19,7 @@ package io.shardingsphere.transaction.bed.sync;
 
 import com.google.common.eventbus.AllowConcurrentEvents;
 import com.google.common.eventbus.Subscribe;
-import io.shardingsphere.core.executor.event.DMLExecutionEvent;
+import io.shardingsphere.core.executor.event.sql.DMLExecutionEvent;
 import io.shardingsphere.transaction.api.SoftTransactionManager;
 import io.shardingsphere.transaction.api.config.SoftTransactionConfiguration;
 import io.shardingsphere.transaction.bed.BEDSoftTransaction;
