+++ sharding-jdbc/src/main/java/io/shardingsphere/core/executor/event/sql/SQLExecutionEventFactory.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2016-2018 shardingsphere.io.
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * </p>
+ */
+
+package io.shardingsphere.core.executor.event.sql;
+
+import io.shardingsphere.core.constant.SQLType;
+import io.shardingsphere.core.executor.BaseStatementUnit;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+/**
+ * SQL execution event.
+ * 
+ * @author gaohongtao
+ * @author maxiaoguang
+ */
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class SQLExecutionEventFactory {
+    
+    /**
+     * Create SQL execution event.
+     * 
+     * @param sqlType SQL type
+     * @param baseStatementUnit statement unit
+     * @param parameters parameters
+     * @return SQL execution event
+     */
+    public static SQLExecutionEvent createEvent(final SQLType sqlType, final BaseStatementUnit baseStatementUnit, final List<Object> parameters) {
+        if (SQLType.DQL == sqlType) {
+            return new DQLExecutionEvent(baseStatementUnit.getSqlExecutionUnit().getDataSource(), baseStatementUnit.getSqlExecutionUnit().getSqlUnit(), parameters);
+        }
+        if (SQLType.DML == sqlType) {
+            return new DMLExecutionEvent(baseStatementUnit.getSqlExecutionUnit().getDataSource(), baseStatementUnit.getSqlExecutionUnit().getSqlUnit(), parameters);
+        }
+        return new SQLExecutionEvent(baseStatementUnit.getSqlExecutionUnit().getDataSource(), baseStatementUnit.getSqlExecutionUnit().getSqlUnit(), parameters);
+    }
+}
