+++ sharding-jdbc/src/test/java/io/shardingsphere/core/executor/fixture/TestDMLExecutionEventListener.java
@@ -19,7 +19,7 @@ package io.shardingsphere.core.executor.fixture;
 
 import com.google.common.eventbus.AllowConcurrentEvents;
 import com.google.common.eventbus.Subscribe;
-import io.shardingsphere.core.executor.event.DMLExecutionEvent;
+import io.shardingsphere.core.executor.event.sql.DMLExecutionEvent;
 import lombok.AllArgsConstructor;
 
 @AllArgsConstructor
