+++ sharding-jdbc/src/test/java/io/shardingsphere/core/executor/fixture/TestOverallExecutionEventListener.java
@@ -19,7 +19,7 @@ package io.shardingsphere.core.executor.fixture;
 
 import com.google.common.eventbus.AllowConcurrentEvents;
 import com.google.common.eventbus.Subscribe;
-import io.shardingsphere.core.executor.event.OverallExecutionEvent;
+import io.shardingsphere.core.executor.event.overall.OverallExecutionEvent;
 import lombok.AllArgsConstructor;
 
 @AllArgsConstructor
